+++ controller/konnect/konnectextension_controller_rbac.go
@@ -7,3 +7,5 @@ package konnect
 // +kubebuilder:rbac:groups=gateway-operator.konghq.com,resources=dataplanes,verbs=get;list;watch
 // +kubebuilder:rbac:groups=gateway-operator.konghq.com,resources=konnectextensions,verbs=get;list;watch
 // +kubebuilder:rbac:groups=gateway-operator.konghq.com,resources=konnectextensions/finalizers,verbs=update;patch
+
+// +kubebuilder:rbac:groups="",resources=secrets,verbs=get;list;watch