+++ CHANGELOG.md
@@ -92,6 +92,10 @@
 - Added support for `KonnectCloudGatewayNetwork` CRD which can manage Konnect
   Cloud Gateway Network entities.
   [#1136](https://github.com/Kong/gateway-operator/pull/1136)
+- Reconcile affected `KonnectExtension`s when the `Secret` used as Dataplane
+  certificate is modified. A secret must have the `konghq.com/konnect-dp-cert`
+  label to trigger the reconciliation.
+  [#1250](https://github.com/Kong/gateway-operator/pull/1250)
 - When the `DataPlane` is configured in Konnect, the `/status/ready` endpoint
   is set as the readiness probe.
   [#1235](https://github.com/Kong/gateway-operator/pull/1253)